+++ discovery-provider/src/queries/search_track_tags.py
@@ -3,7 +3,8 @@ import logging  # pylint: disable=C0302
 from src.models.tracks.tag_track_user_matview import t_tag_track_user
 from src.models.tracks.track import Track
 from src.queries import response_name_constants
-from src.queries.query_helpers import get_track_play_counts, populate_track_metadata
+from src.queries.query_helpers import populate_track_metadata
+from src.queries.sol_play_helpers import get_track_play_counts
 from src.utils import helpers
 
 logger = logging.getLogger(__name__)
